Fixed some issues with how we import from .txt
First of all there was an issue with the path of the files. Just added the folder before "src". Think it fixed it

Second, there was quite a bit of problems and mistakes with how we imported and created instances of the different objects from a file. That should be fixed

Fixed the issue with us not being able to watch credits from both a transmission and an episode of a tv-show -

Added a few comments

Cleaned up the SystemAdminController