Updated createdBy and the LogIn functionality
THE IMPORT AND EXPORT WORKS!!!!
Know that it is going to be removed next week but needed it for testing other functionality

Changed how we store and read createdBy. Now as a UUID

The logIn functionality does not skip a layer and the logInHandler has been moved out of the accesscontroller-folder to the domain-folder. Both the logInHandler and the User file has been updated

Changed primary.fxml to logInPage.fxml

Removed UUID and CreatedBy from guest view. That created another issue; we cannot get credits when viewing as a guest because we do not have any ID to match the program and get the credits on that current program

Updated the approve-functionality. Now when you update wither a program or a credit an admin has to approve once again.

Changed createdBy showing as UUID from the listView to the name of the user

Confirmed that only the person who created the program can update it no matter what - made sure that when an admin updated a program it doesn't change the "createdBy" attribute

Made a separate method for the approved-tab to get the program (Didn't remove the old method in case  we want to fix it)

In the accesscontroller-package I made the systemAdmin to extend from user and not producer.